scripts/requirements.txt
pyyaml>=6.0.2
pydantic>=2.8
typer>=0.12
rich>=13.7
pyyaml
numpy>=1.26
pandas>=2.2
mkdocs>=1.5.3
mkdocs-material>=9.5
scikit-learn>=1.4